+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.0] - 2019-04-19
 ### Added
 - Replace `def foo=(value); @foo = value; end` with `attr_writer :foo`
 - Replace `while false ... end` loops with nothing
@@ -34,7 +36,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 - Initial release. 🎉
 
-[Unreleased]: https://github.com/kddeisz/preval/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/kddeisz/preval/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/kddeisz/preval/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/kddeisz/preval/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/kddeisz/preval/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/kddeisz/preval/compare/49c899...v0.1.0